// packages/container-bi-query/lib/config/types.ts
export enum FieldName {
  /// type1 
  actionDay = 'action_day',
  actionHour = 'action_hour',
  agentName = 'agent_name',
  customerName = 'customer_name',
  projectName = 'project_name',
  campaignName = 'campaign_name',
  contractName = 'contract_name',
  orderName = 'order_name',
  creativeName = 'creative_name',
  frequency = 'frequency',
  adxId = 'adx_id',
  adxName = 'adx_name',
  requestTypeName = 'request_type_name',
  requestTerminal = 'request_terminal',
  displayForm = 'display_form',
  relevantType = 'relevant_type',
  dealId = 'deal_id',
  appId = 'app_id',
  appPackage = 'app_package',
  appName = 'app_name',
  appCategoryName = 'app_category_name',
  country = 'country',
  province = 'province',
  city = 'city',
  cityLevel = 'city_level',
  county = 'county',
  agentId = 'agent_id',
  customerId = 'customer_id',
  projectId = 'project_id',
  contractId = 'contract_id',
  campaignId = 'campaign_id',
  orderId = 'order_id',
  creativeId = 'creative_id',
  adId = 'ad_id',
  creativeDynamicType = 'creative_dynamic_type',
  otvPosition = 'otv_position',
  adzoneId = 'adzone_id',
  adzoneName = 'adzone_name',
  adzoneSize = 'adzone_size',
  requestMaxDuration = 'request_max_duration',
  requestMinDuration = 'request_min_duration',
  creativeVideoDuration = 'creative_video_duration',
  nativeId = 'native_id',
  nativeName = 'native_name',
  bidNativeId = 'bid_native_id',
  bidNativeName = 'bid_native_name',
  os = 'os',
  channelName = 'channel_name',
  deviceBrand = 'device_brand',
  conversionName = 'conversion_name',
  conversionOrderId = 'conversion_order_id',
  conversionDetail = 'conversion_detail',
  conversionIdentity = 'conversion_identity',
  impressionTime = 'impression_time',
  clickTime = 'click_time',
  conversionTime = 'conversion_time',
  adzonePosition = 'adzone_position',
  mainDomain = 'main_domain',
  effectDelayDay = 'effect_delay_day',
  purchaseFromType = 'purchase_from_type',
  adzoneReservePrice = 'adzone_reserve_price',
  ask = 'ask',
  bid = 'bid',
  win = 'win',
  impression = 'impression',
  click = 'click',
  reach = 'reach',
  conversion = 'conversion',
  impConversion = 'imp_conversion',
  clickConversion = 'click_conversion',
  askUv = 'ask_uv',
  bidUv = 'bid_uv',
  winUv = 'win_uv',
  impressionUv = 'impression_uv',
  clickUv = 'click_uv',
  reachUv = 'reach_uv',
  conversionUv = 'conversion_uv',
  btr = 'btr',
  wtr = 'wtr',
  ctr = 'ctr',
  ecpm = 'ecpm',
  agentEcpm = 'agent_ecpm',
  customerEcpm = 'customer_ecpm',
  cpc = 'cpc',
  agentCpc = 'agent_cpc',
  customerCpc = 'customer_cpc',
  cost = 'cost',
  agentCost = 'agent_cost',
  customerCost = 'customer_cost',
  contractCost = 'contract_cost',
  lostCost = 'lost_cost',
  conversionPrice = 'conversion_price',
  deepLinkAccess = 'deep_link_access',
  deepLinkEvoke = 'deep_link_evoke',
  deepLinkEvokeSuccess = 'deep_link_evoke_success',
  deepLinkEvokeFail = 'deep_link_evoke_fail',

  week = 'week',
  uv = 'uv',
  hourOfWeek = 'hour_of_week',
  deviceTypeWeek = 'device_type_week',
  adFormatWeek = 'ad_format_week',
  provinceWeek = 'province_week',
}